Creation of the store filter
 I extracted the filter logic from the store program.
 I also adjusted the imports to use the one from the filter.
