+++ packages/web/src/stores/filter.ts
@@ -1,0 +1,21 @@
+import { defineStore } from 'pinia'
+import { type programFiltersType } from '@/types'
+
+export const useFilterStore = defineStore('filter', () => {
+  const programFilters = ref<programFiltersType>({
+    programAidTypeSelected: '',
+    objectifTypeSelected: ''
+  })
+
+  function resetFilters() {
+    programFilters.value = {
+      programAidTypeSelected: '',
+      objectifTypeSelected: ''
+    }
+  }
+
+  return {
+    resetFilters,
+    programFilters
+  }
+})
